Rename chart directories
