+++ src/routes/components/charts/common/chart-datum.ts
