+++ src/routes/components/charts/common/index.ts
@@ -1,1 +1,1 @@
-export { ChartDatum } from './chart-datum-utils';
+export { ChartDatum } from './chart-datum';
