+++ src/routes/components/charts/trend/TrendChart.styles.ts
