+++ src/routes/components/charts/trend/index.ts
